+++ sample/Frank.OpenApi.Sample/Extensions.fs
@@ -1,0 +1,24 @@
+module Sample.OpenApi.Extensions
+
+open Microsoft.AspNetCore.Builder
+open Microsoft.AspNetCore.Hosting
+open Microsoft.Extensions.DependencyInjection
+open Microsoft.Extensions.Hosting
+open Microsoft.Extensions.Logging
+
+/// Helper function to determine whether the application
+/// is running in a development environment.
+let isDevelopment (app:IApplicationBuilder) =
+    app.ApplicationServices
+       .GetService<IWebHostEnvironment>()
+       .IsDevelopment()
+
+type Frank.Builder.WebHostBuilder with
+
+    /// Extension to the WebHostBuilder computation expression
+    /// to support logging through the ILoggingBuilder.
+    [<CustomOperation("logging")>]
+    member __.Logging(spec:Frank.Builder.WebHostSpec, f:ILoggingBuilder -> ILoggingBuilder) =
+        { spec with
+            Services = fun services ->
+                spec.Services(services).AddLogging(fun builder -> f builder |> ignore) }